requirements.txt
autocorrect
dateparser==0.7.6
fuzzywuzzy==0.18.0
googletrans==3.0.0
nltk==3.5
numpy==1.23.1
pandas==1.4.3
pyspellchecker==0.5.4
python-Levenshtein==0.12.0
requests==2.22.0
sqlalchemy
Shapely==1.8.2
